+++ modules/examples/example-tf-prereqs/stack.py
@@ -29,6 +29,7 @@ class TfPreReqs(Stack):  # type: ignore
         self,
         scope: Construct,
         id: str,
+        project_name: str,
         deployment_name: str,
         module_name: str,
         hash: str,
@@ -40,12 +41,10 @@ class TfPreReqs(Stack):  # type: ignore
         super().__init__(
             scope,
             id,
-            description="This stack deploys Storage resources for ADDF",
+            description="This stack deploys Storage resources",
             **kwargs,
         )
-        Tags.of(scope=cast(IConstruct, self)).add(
-            key="Deployment", value=f"addf-{deployment_name}"
-        )
+        Tags.of(scope=cast(IConstruct, self)).add(key="Deployment", value=f"{project_name}-{deployment_name}")
 
         # S3 bucket for storing the remote state of Terraform
         self.tf_state_s3bucket = aws_s3.Bucket(
@@ -53,7 +52,7 @@ class TfPreReqs(Stack):  # type: ignore
             removal_policy=aws_cdk.RemovalPolicy.RETAIN
             if tf_s3_backend_retention_type.upper() == "RETAIN"
             else aws_cdk.RemovalPolicy.DESTROY,
-            bucket_name=f"addf-{deployment_name}-tfstate-bucket-{hash}",
+            bucket_name=f"{project_name}-{deployment_name}-tfstate-bucket-{hash}",
             id="tf-state-bucket",
             encryption=aws_s3.BucketEncryption.KMS_MANAGED
             if tf_s3_backend_encryption_type.upper() == "KMS"
@@ -68,10 +67,8 @@ class TfPreReqs(Stack):  # type: ignore
         self.tf_ddb_lock_table = aws_dynamodb.Table(
             self,
             "tf_ddb_lock_table",
-            table_name=f"addf-{deployment_name}-tf-ddb-lock-table",
-            partition_key=aws_dynamodb.Attribute(
-                name=part_key, type=aws_dynamodb.AttributeType.STRING
-            ),
+            table_name=f"{project_name}-{deployment_name}-tf-ddb-lock-table",
+            partition_key=aws_dynamodb.Attribute(name=part_key, type=aws_dynamodb.AttributeType.STRING),
             billing_mode=aws_dynamodb.BillingMode.PAY_PER_REQUEST,
             removal_policy=aws_cdk.RemovalPolicy.RETAIN
             if tf_ddb_retention_type.upper() == "RETAIN"