+++ modules/visualization/foxglove-image/app.py
@@ -14,7 +14,9 @@ deployment_name = os.getenv("SEEDFARMER_DEPLOYMENT_NAME")
 module_name = os.getenv("SEEDFARMER_MODULE_NAME")
 
 if len(f"{project_name}-{deployment_name}") > 36:
-    raise ValueError("This module cannot support a project+deployment name character length greater than 35")
+    raise ValueError(
+        "This module cannot support a project+deployment name character length greater than 35"
+    )
 
 
 def _param(name: str) -> str: